pyproject.toml
[project]
name = "elevenlabs-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs>=2.40.0",
    "legacy-cgi>=2.6.4",
    "python-dotenv>=1.2.2",
    "streamlit>=1.28.0",
]
